Add Incident Map Page (#16)
* Add Incident Map Page

* Fix histogram bins

Histogram fixed so that columns line up with appropriate year. Previously they were one year off.

* Incident map alpha release

* Fix count graph

With the larger dataset and changes to the database schema, the code needed to be updated to match.

* Correct map load error

Corrected an error in which the map failed to generate due to a data structure error.