Remove hidden div with full database
Remove the hidden div that stores the full json form of the database on page load. Instead place a query function inside each callback that requires the data (map, summary boxes, count graph, description). This is slower, but it prevents potential scraping of the protected data. Callbacks execute on the server side, meaning the full data never sees the end user's system.